testing failure not working

+++ src/main/java/la/crdt/TimeStamp.java
@@ -4,9 +4,9 @@ import java.io.Serializable;
 public class TimeStamp implements Serializable{
 	private static final long serialVersionUID = 11L;
 	public int id;
-	public int clock;
+	public long clock;
 
-	public TimeStamp(int id, int clock) {
+	public TimeStamp(int id, long clock) {
 		this.id = id;
 		this.clock = clock;
 	}
